Move logic of adjust threshold to zmalloc module
