+++ src/server.c
@@ -1697,9 +1697,6 @@ void databasesCron(void) {
         }
     }
 
-    /* Adjust PMEM threshold. */
-    adjustPmemThresholdCycle();
-
     /* Defrag keys gradually. */
     activeDefragCycle();
 
@@ -2736,6 +2733,28 @@ void resetServerStats(void) {
     server.stat_net_output_bytes = 0;
     server.stat_unexpected_error_replies = 0;
     server.aof_delayed_fsync = 0;
+}
+
+/* Initialize the pmem threshold. */
+static void pmemThresholdInit(void) {
+    switch(server.memory_alloc_policy) {
+        case MEM_POLICY_ONLY_DRAM:
+            zmalloc_set_threshold(UINT_MAX);
+            break;
+        case MEM_POLICY_ONLY_PMEM:
+            zmalloc_set_threshold(0U);
+            break;
+        case MEM_POLICY_THRESHOLD:
+            zmalloc_set_threshold(server.static_threshold);
+            break;
+        case MEM_POLICY_RATIO:
+            zmalloc_set_threshold(server.initial_dynamic_threshold);
+            zmalloc_set_ratio(server.target_pmem_dram_ratio);
+            zmalloc_set_threshold_range(server.dynamic_threshold_min, server.dynamic_threshold_max);
+            break;
+        default:
+            serverAssert(NULL);
+    }
 }
 
 void initServer(void) {
