+++ tests/unit/introspection.tcl
@@ -103,7 +103,6 @@ start_server {tags {"introspection"}} {
             initial-dynamic-threshold
             dynamic-threshold-min
             dynamic-threshold-max
-            memory-ratio-check-period
             hashtable-on-dram
         }
